ball: Clean up ball manager.
