+++ VisualPinball.Unity/VisualPinball.Unity/Physics/Engine/DefaultPhysicsEngine.cs
@@ -28,6 +28,7 @@ namespace VisualPinball.Unity
 	{
 		public string Name => "Default VPX";
 
+		private BallManager _ballManager;
 		private EntityManager _entityManager;
 		private EntityQuery _flipperDataQuery;
 		private EntityQuery _ballDataQuery;
@@ -46,6 +47,7 @@ namespace VisualPinball.Unity
 				ComponentType.ReadOnly<SolenoidStateData>()
 			);
 
+			_ballManager = BallManager.Instance(tableAuthoring.Table, tableAuthoring.gameObject.transform.localToWorldMatrix);
 			_ballDataQuery = _entityManager.CreateEntityQuery(ComponentType.ReadOnly<BallData>());
 
 			var visualPinballSimulationSystemGroup = _entityManager.World.GetOrCreateSystem<VisualPinballSimulationSystemGroup>();
@@ -60,7 +62,7 @@ namespace VisualPinball.Unity
 		public void BallCreate(Mesh mesh, Material material, in float3 worldPos, in float3 localPos,
 			in float3 localVel, in float scale, in float mass, in float radius, in Entity kickerRef)
 		{
-			BallManager.CreateEntity(mesh, material, in worldPos, in localPos, in localVel,
+			_ballManager.CreateEntity(mesh, material, in worldPos, in localPos, in localVel,
 				scale * radius * 2, in mass, in radius, in kickerRef);
 		}
 
@@ -149,11 +151,11 @@ namespace VisualPinball.Unity
 
 		public void PushPendingCreateBallNotifications()
 		{
-			if (_nextBallIdToNotifyDebugUI == BallAuthoring.NumBallsCreated)
+			if (_nextBallIdToNotifyDebugUI == BallManager.NumBallsCreated)
 				return; // nothing to report
 
 			var entities = _ballDataQuery.ToEntityArray(Allocator.TempJob);
-			int numBallsToReport = BallAuthoring.NumBallsCreated - _nextBallIdToNotifyDebugUI;
+			int numBallsToReport = BallManager.NumBallsCreated - _nextBallIdToNotifyDebugUI;
 			foreach (var entity in entities)
 			{
 				var ballData = _entityManager.GetComponentData<BallData>(entity);
@@ -166,7 +168,7 @@ namespace VisualPinball.Unity
 
 			// error checking
 			Assert.AreEqual(0, numBallsToReport);
-			_nextBallIdToNotifyDebugUI = BallAuthoring.NumBallsCreated;
+			_nextBallIdToNotifyDebugUI = BallManager.NumBallsCreated;
 		}
 	}
 }
