--- VisualPinball.Unity/VisualPinball.Unity/VPT/Ball/BallAuthoring.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b9b9edb75ce238a4f93855b729e8cebf
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
