+++ VisualPinball.Unity/VisualPinball.Unity/VPT/Ball/BallManager.cs
@@ -22,52 +22,148 @@ using UnityEngine;
 using UnityEngine.Rendering;
 using VisualPinball.Engine.Common;
 using VisualPinball.Engine.Game;
+using VisualPinball.Engine.VPT.Table;
 
 namespace VisualPinball.Unity
 {
+	/// <summary>
+	/// A singleton class that handles ball creation and destruction.
+	/// </summary>
 	internal class BallManager
 	{
-		private int _id;
+		public static int NumBallsCreated { get; private set; }
 
-		private readonly Engine.VPT.Table.Table _table;
+		private readonly Table _table;
+		private readonly Matrix4x4 _ltw;
+		private readonly EntityManager _entityManager;
 
+		private static BallManager _instance;
 		private static Mesh _unitySphereMesh; // used to cache ball mesh from GameObject
 
-		public BallManager(Engine.VPT.Table.Table table)
+		public static BallManager Instance(Table table, Matrix4x4 ltw) => _instance ?? (_instance = new BallManager(table, ltw));
+
+		public BallManager(Table table, Matrix4x4 ltw)
 		{
 			_table = table;
+			_ltw = ltw;
+			_entityManager = World.DefaultGameObjectInjectionWorld.EntityManager;
 		}
 
-		public void CreateBall(Player player, IBallCreationPosition ballCreator, float radius, float mass, in Entity kickerRef)
+		public void CreateBall(IBallCreationPosition ballCreator, float radius = 25f, float mass = 1f)
 		{
-			// calculate mass and scale
-			var ltw = player.TableToWorld;
+			CreateBall(ballCreator, radius, mass, Entity.Null);
+		}
 
+		public void CreateBall(IBallCreationPosition ballCreator, float radius, float mass, in Entity kickerRef)
+		{
 			var localPos = ballCreator.GetBallCreationPosition(_table).ToUnityFloat3();
 			var localVel = ballCreator.GetBallCreationVelocity(_table).ToUnityFloat3();
 			localPos.z += radius;
-			//float4x4 model = player.TableToWorld * Matrix4x4.TRS(localPos, Quaternion.identity, new float3(radius));
 
-			var worldPos = ltw.MultiplyPoint(localPos);
+			var worldPos = _ltw.MultiplyPoint(localPos);
 			var scale3 = new Vector3(
-				ltw.GetColumn(0).magnitude,
-				ltw.GetColumn(1).magnitude,
-				ltw.GetColumn(2).magnitude
+				_ltw.GetColumn(0).magnitude,
+				_ltw.GetColumn(1).magnitude,
+				_ltw.GetColumn(2).magnitude
 			);
 			var scale = (scale3.x + scale3.y + scale3.z) / 3.0f; // scale is only scale (without radiusfloat now, not vector.
 			var material = BallMaterial.CreateMaterial();
 			var mesh = GetSphereMesh();
 
 			// create ball entity
-			EngineProvider<IPhysicsEngine>.Get()
+			EngineProvider<IPhysicsEngine>
+				.Get()
 				.BallCreate(mesh, material, worldPos, localPos, localVel, scale, mass, radius, in kickerRef);
 		}
 
-		public static void CreateEntity(Mesh mesh, Material material, in float3 worldPos, in float3 localPos,
+		public void CreateEntity(Mesh mesh, Material material, in float3 worldPos, in float3 localPos,
 			in float3 localVel, in float scale, in float mass, in float radius, in Entity kickerEntity)
 		{
-			var entityManager = World.DefaultGameObjectInjectionWorld.EntityManager;
-			BallAuthoring.CreateEntity(entityManager, mesh, material, worldPos, scale, localPos, localVel, radius, mass, kickerEntity);
+			var world = World.DefaultGameObjectInjectionWorld;
+			var ecbs = world.GetOrCreateSystem<CreateBallEntityCommandBufferSystem>();
+			var ecb = ecbs.CreateCommandBuffer();
+
+			var archetype = _entityManager.CreateArchetype(
+				typeof(Translation),
+				typeof(Scale),
+				typeof(Rotation),
+				typeof(RenderMesh),
+				typeof(LocalToWorld),
+				typeof(RenderBounds),
+				typeof(BallData),
+				typeof(CollisionEventData),
+				typeof(OverlappingStaticColliderBufferElement),
+				typeof(OverlappingDynamicBufferElement),
+				typeof(ContactBufferElement),
+				typeof(BallInsideOfBufferElement),
+				typeof(BallLastPositionsBufferElement)
+			);
+			var entity = ecb.CreateEntity(archetype);
+
+			ecb.SetSharedComponent(entity, new RenderMesh {
+				mesh = mesh,
+				material = material
+			});
+
+			ecb.SetComponent(entity, new RenderBounds {
+				Value = mesh.bounds.ToAABB()
+			});
+
+			ecb.SetComponent(entity, new Translation {
+				Value = worldPos
+			});
+
+			ecb.SetComponent(entity, new Scale {
+				Value = scale
+			});
+
+			ecb.SetComponent(entity, new BallData {
+				Id = NumBallsCreated++,
+				IsFrozen = false,
+				Position = localPos,
+				Radius = radius,
+				Mass = mass,
+				Velocity = localVel,
+				Orientation = float3x3.identity,
+				RingCounterOldPos = 0,
+				AngularMomentum = float3.zero
+			});
+
+			ecb.SetComponent(entity, new CollisionEventData {
+				HitTime = -1,
+				HitDistance = 0,
+				HitFlag = false,
+				IsContact = false,
+				HitNormal = new float3(0, 0, 0),
+			});
+
+			var lastBallPostBuffer = ecb.AddBuffer<BallLastPositionsBufferElement>(entity);
+			for (var i = 0; i < BallRingCounterSystem.MaxBallTrailPos; i++) {
+				lastBallPostBuffer.Add(new BallLastPositionsBufferElement
+					{ Value = new float3(float.MaxValue, float.MaxValue, float.MaxValue) }
+				);
+			}
+
+			// handle inside-kicker creation
+			if (kickerEntity != Entity.Null) {
+				var kickerData = _entityManager.GetComponentData<KickerStaticData>(kickerEntity);
+				if (!kickerData.FallThrough) {
+					var kickerCollData = _entityManager.GetComponentData<KickerCollisionData>(kickerEntity);
+					var inside = ecb.AddBuffer<BallInsideOfBufferElement>(entity);
+					BallData.SetInsideOf(ref inside, kickerEntity);
+					kickerCollData.BallEntity = entity;
+					kickerCollData.LastCapturedBallEntity = entity;
+					ecb.SetComponent(kickerEntity, kickerCollData);
+				}
+			}
+		}
+
+		public static void DestroyEntity(Entity ballEntity)
+		{
+			World.DefaultGameObjectInjectionWorld
+				.GetOrCreateSystem<CreateBallEntityCommandBufferSystem>()
+				.CreateCommandBuffer()
+				.DestroyEntity(ballEntity);
 		}
 
 		/// <summary>
@@ -87,25 +183,6 @@ namespace VisualPinball.Unity
 			return _unitySphereMesh;
 		}
 
-		private GameObject CreateSphere(Material material, float3 pos, float3 scale)
-		{
-			// create go
-			var go = GameObject.CreatePrimitive(PrimitiveType.Sphere);
-			go.name = $"Ball{++_id}";
-
-			// set material
-			go.GetComponent<Renderer>().material = material;
-
-			// set position and scale
-			go.transform.localPosition = pos;
-			go.transform.localScale = scale;
-
-			// mark to convert
-			go.AddComponent<ConvertToEntity>();
-
-			return go;
-		}
-
 		#region Material
 
 		/// <summary>
@@ -119,14 +196,12 @@ namespace VisualPinball.Unity
 		/// <returns></returns>
 		private static IBallMaterial CreateBallMaterial()
 		{
-			if (GraphicsSettings.renderPipelineAsset != null)
-			{
-				if (GraphicsSettings.renderPipelineAsset.GetType().Name.Contains("UniversalRenderPipelineAsset"))
-				{
+			if (GraphicsSettings.renderPipelineAsset != null) {
+				if (GraphicsSettings.renderPipelineAsset.GetType().Name.Contains("UniversalRenderPipelineAsset")) {
 					return new UrpBallMaterial();
 				}
-				else if (GraphicsSettings.renderPipelineAsset.GetType().Name.Contains("HDRenderPipelineAsset"))
-				{
+
+				if (GraphicsSettings.renderPipelineAsset.GetType().Name.Contains("HDRenderPipelineAsset")) {
 					return new HdrpBallMaterial();
 				}
 			}
@@ -135,6 +210,5 @@ namespace VisualPinball.Unity
 		}
 
 		#endregion
-
 	}
 }
